Legacy_Code_Conv/Images/Modernization/requirements.txt
# Core dependencies
fastapi>=0.104.1
uvicorn>=0.24.0
python-multipart>=0.0.6
pydantic>=2.5.2
pydantic-settings>=2.1.0

# Image processing
Pillow>=10.1.0
python-magic>=0.4.27

# Storage and caching
boto3>=1.29.3
redis>=5.0.1
aiofiles>=23.2.1
aiohttp>=3.9.1

# Security
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4

# Testing
pytest>=7.4.3
pytest-asyncio>=0.21.1
pytest-cov>=4.1.0
httpx>=0.25.2

# Development
black>=23.11.0
isort>=5.12.0
mypy>=1.7.1
flake8>=6.1.0
